requirements.txt
torch==2.5.1
torchvision==0.20.1

opencv-python==4.10.0.84
omegaconf==2.3.0
tqdm==4.67.1
pillow==11.0.0
natsort==8.4.0
pycocotools==2.0.8
diffusers==0.31.0
transformers==4.47.0
einops==0.8.0
accelerate==1.2.0
peft==0.14.0
scipy==1.14.1
shapely==2.0.6
scikit-image==0.24.0
